Apply black formatting

+++ Test/Models/test_autoreg.py
@@ -26,24 +26,19 @@ def test_ARNNDense():
     model = nk.models.ARNNDense(layers=3, features=5)
     in_shape = (batch, size)
     key_spins, key_model = jax.random.split(jax.random.PRNGKey(0))
-    spins = jax.random.bernoulli(key_spins, shape=in_shape).astype(
-        model.dtype) * 2 - 1
-    (p, _), params = model.init_with_output(key_model,
-                                            spins,
-                                            None,
-                                            method=model.conditionals)
+    spins = jax.random.bernoulli(key_spins, shape=in_shape).astype(model.dtype) * 2 - 1
+    (p, _), params = model.init_with_output(
+        key_model, spins, None, method=model.conditionals
+    )
 
     for i in range(batch):
         for j in range(size):
             # Change one input element at a time
             spins_new = spins.at[i, j].set(-spins[i, j])
-            p_new, _ = model.apply(params,
-                                   spins_new,
-                                   None,
-                                   method=model.conditionals)
+            p_new, _ = model.apply(params, spins_new, None, method=model.conditionals)
             p_diff = p_new - p
 
             # The former output elements should not change
-            p_diff = p_diff.at[i, j + 1:].set(0)
+            p_diff = p_diff.at[i, j + 1 :].set(0)
 
-            np.testing.assert_allclose(p_diff, 0, err_msg=f'{i=} {j=}')
+            np.testing.assert_allclose(p_diff, 0, err_msg=f"{i=} {j=}")
